Email Msg on Whatsapp/requirements.txt
twilio
python-dotenv
imaplib2
